Move SeqMems to BigInts and plumb through test mems (#1934)
Enables larger ExtMem sizes with the current TestHarness.

+++ src/main/scala/diplomaticobjectmodel/DiplomaticObjectModelUtils.scala
@@ -181,7 +181,7 @@ object DiplomaticObjectModelAddressing {
 
   def makeOMMemory[T <: Data](
       desc: String,
-      depth: Int,
+      depth: BigInt,
       data: T
     ): OMMemory = {
